% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{anno}
\alias{anno}
\title{Annotation with MAIT package}
\usage{
anno(path, name, mode = NULL)
}
\arguments{
\item{path}{the path to your data}

\item{name}{the name of your project}

\item{mode}{defalut is for positive mode, otherwise use 'negAdducts' for negative}
}
\value{
as shown in MAIT package
}
\description{
Annotation with MAIT package
}
\examples{
\dontrun{
path <- "./data/"
name <- "fishriver"
anno(path,name)
}
}
\references{
Fernández-Albert, F.; Llorach, R.; Andrés-Lacueva, C.; Perera, A. Bioinformatics 2014, 30 (13), 1937–1939.
}
