% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{fanno2}
\alias{fanno2}
\title{Annotation with xMSannotator package for list}
\usage{
fanno2(list, outloc = "./result/", mode = "pos", db_name = "HMDB",
  num_nodes = 10, ppm = 5, ...)
}
\arguments{
\item{list}{a list to be annotated}

\item{outloc}{the path for your result}

\item{mode}{defalut is for positive mode, otherwise use 'neg' for negative}

\item{db_name}{default is 'HMDB', other database options: 'KEGG', 'LipidMaps', 'T3DB'}

\item{num_nodes}{default 10}

\item{ppm}{for mass accuracy}

\item{...}{parameters for multilevelannotation function in xMSannotator}
}
\value{
as shown in xMSannotator package
}
\description{
Annotation with xMSannotator package for list
}
\examples{
\dontrun{
path <- "./data/"
xset <- getdata(path)
list <- getmzrt(xset)
result <- fanno2(list)
}
}
\references{
Uppal, K.; Walker, D. I.; Jones, D. P. Anal. Chem. 2017, 89 (2), 1063–1067.
}
