% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplus.R
\name{getprop}
\alias{getprop}
\title{Function for Good/Bad peaks balance}
\usage{
getprop(xset, cutoff = 10)
}
\arguments{
\item{xset}{xcmsSet object}

\item{cutoff}{minium point numbers for good peaks}
}
\value{
Good peaks proportion and bad peaks proportion
}
\description{
Function for Good/Bad peaks balance
}
