% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{mumdata}
\alias{mumdata}
\title{Output pathway annotation data for Mummichog algorithm}
\usage{
mumdata(xset, lv = NULL, name = "test", method = "medret",
  intensity = "into")
}
\arguments{
\item{xset}{a xcmsset object to be annotated}

\item{lv}{the group factor}

\item{name}{the file name}

\item{method}{parameter for groupval function}

\item{intensity}{parameter for groupval function}
}
\value{
text file for Mummichog algorithm
}
\description{
Output pathway annotation data for Mummichog algorithm
}
\details{
use 'python2 main.py -c 0.05 -f test.txt -p 100 -m negative -o myoutput' to get the Mummichog annotation
}
\references{
Li, S.; Park, Y.; Duraisingham, S.; Strobel, F. H.; Khan, N.; Soltow, Q. A.; Jones, D. P.; Pulendran, B. PLOS Computational Biology 2013, 9 (7), e1003123.
}
