% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{pmdtarget}
\alias{pmdtarget}
\title{Output pmd results into csv(s) for MS/MS validation}
\usage{
pmdtarget(list, Dppm = 20, Drt = 0.5, ce = NA, name = "target",
  n = NULL)
}
\arguments{
\item{list}{list from getstd or globalstd with independant peaks index}

\item{Dppm}{ppm for targeted analysis}

\item{Drt}{retention time drift for targeted analysis}

\item{ce}{collision energy for MS/MS analysis}

\item{name}{file name}

\item{n}{max ions numbers within retention time drift windows}
}
\description{
Output pmd results into csv(s) for MS/MS validation
}
