% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rtmzsim}
\alias{rtmzsim}
\title{Generate simulated count data with batch effects for npeaks}
\usage{
rtmzsim(npeaks = 1000, nspeaks = 50, nbatch = 3, ncond = 2,
  npercond = 10, basemean = 10000, ppstep = 50, bbstep = 2000,
  ccstep = 800, basedisp = 100, bdispstep = 10, swvar = 1000,
  seed = 42)
}
\arguments{
\item{npeaks}{Number of peaks to simulate}

\item{nspeaks}{Number of significant peaks to simulate}

\item{nbatch}{Number of batches to simulate}

\item{ncond}{Number of conditions to simulate}

\item{npercond}{Number of samples per condition per batch to simulate}

\item{basemean}{Base mean}

\item{ppstep}{peak to peak step variation}

\item{bbstep}{Batch to Batch step variation}

\item{ccstep}{Condition to Condition step variation}

\item{basedisp}{Base Dispersion}

\item{bdispstep}{Batch to Batch Dispersion step variation}

\item{swvar}{Sample-wise extra variation}

\item{seed}{Random seed for reproducibility}
}
\value{
rtmz data matrix
}
\description{
Generate simulated count data with batch effects for npeaks
}
\examples{
rtmzsim()
}
