% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{svaanno}
\alias{svaanno}
\title{Annotation with MAIT package of list from svacor function}
\usage{
svaanno(raw, lv, polarity = "positive", projectname = "test")
}
\arguments{
\item{raw}{the list from svacor function}

\item{lv}{group information}

\item{polarity}{defalut is for positive mode, otherwise use 'negative' for negative}

\item{projectname}{the name of your project}
}
\value{
as shown in MAIT package
}
\description{
Annotation with MAIT package of list from svacor function
}
\references{
Fernández-Albert, F.; Llorach, R.; Andrés-Lacueva, C.; Perera, A. Bioinformatics 2014, 30 (13), 1937–1939.
}
