% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{svafanno}
\alias{svafanno}
\title{Annotation with xMSannotator package}
\usage{
svafanno(raw, outloc = "./result/", mode = "pos", db_name = "HMDB", ...)
}
\arguments{
\item{raw}{a xcmsset object to be annotated}

\item{outloc}{the path for your result}

\item{mode}{defalut is for positive mode, otherwise use 'neg' for negative}

\item{db_name}{default is 'HMDB', other database options: 'KEGG', 'LipidMaps', 'T3DB'}

\item{...}{parameters for multilevelannotation function in xMSannotator}
}
\value{
as shown in xMSannotator package
}
\description{
Annotation with xMSannotator package
}
\references{
Uppal, K.; Walker, D. I.; Jones, D. P. Anal. Chem. 2017, 89 (2), 1063–1067.
}
