% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{targetsep}
\alias{targetsep}
\title{Output pmd results into csv(s) for MS/MS validation}
\usage{
targetsep(rt, Drt, n = 6)
}
\arguments{
\item{rt}{retention time vector for peaks}

\item{Drt}{retention time drift for targeted analysis}

\item{n}{max ions numbers within retention time drift windows}
}
\value{
index for each injection
}
\description{
Output pmd results into csv(s) for MS/MS validation
}
