% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{fanno}
\alias{fanno}
\title{Annotation with xMSannotator package}
\usage{
fanno(xset, outloc = "./result/", mode = "pos", db_name = "HMDB",
  num_nodes = 10, ...)
}
\arguments{
\item{xset}{a xcmsset object to be annotated}

\item{outloc}{the path for your result}

\item{mode}{defalut is for positive mode, otherwise use 'neg' for negative}

\item{db_name}{default is 'HMDB', other database options: 'KEGG', 'LipidMaps', 'T3DB'}

\item{...}{parameters for multilevelannotation function in xMSannotator}
}
\value{
as shown in xMSannotator package
}
\description{
Annotation with xMSannotator package
}
\examples{
\dontrun{
path <- "./data/"
xset <- getdata(path)
result <- fanno(xset)
}
}
\references{
Uppal, K.; Walker, D. I.; Jones, D. P. Anal. Chem. 2017, 89 (2), 1063–1067.
}
